refactor(subscriptions): Support returning either side of a join (#639)
* refactor(subscriptions): Support returning either side of a join

Before this patch we made reference to the "lhs" and the "rhs" of an index join,
in the context of incremental evaluation.
This patch further generalizes the incremental evaluation of index joins,
by replacing such references with either the probe side or the index side.
This is in preparation for adding more types of incremental evaluation plans.

* test: Move incremental eval compilation tests to subscription module