`MutTxDatastore::table_name_from_id_mut_tx`: return `Cow<str>` (#700)
* `MutTxDatastore::table_name_from_id_mut_tx`: return `Cow<str>`

This commit changes the `MutTxDatastore` trait's method `table_name_from_id_mut_tx` method
to return a `Cow<str>` rather than an `&str`,
abstracting away whether the underlying datastore stores a contiguous `str` table name.

As with PR #691, our current datastore has a `String` to which it can return an `&str`,
but the mem-arch-prototype stores a new representation
where strings are not necessarily contiguous,
so it must extract the table name and return an owned `String` rather than an `&str`.

* Rewrite a horrible memoizing branch into the entry API