Integrate mem arch into `locking_tx_datastore` (#769)
* datastore: impl mem arch using spacetimedb_table

* instance_env/iters: write bflatn -> bsatn directly

* Make tests build, fix 1 failure (of 4)

Commit message from SpacetimeDBPrivate 0ae26519d181cb4b13f748af53d036c322d4f642:

Bug fix in `delete_by_rel`: don't trigger unique constraints in dummy insert

Prior to this commit, `MutTxId::delete_by_rel` on a table with a unique constraint
would fail, because the `delete_by_rel` would attempt an insert of the row to be deleted,
with the intention of rolling back the insertion later.
This would trigger the unique constraint, aborting the `delete_by_rel`,
and the actual row would not be deleted.

With this commit, `MutTxId::delete_by_rel` properly uses
`Table::insert_internal_allow_duplicate` and `Table::delete_internal_skip_pointer_map`
to bypass unique constraints.
This requires exporting some methods from `Table` which were previously internal.

* Remove forgotten `dbg`

* `bflatn_to`: panic rather than UB on an ill-typed `ProductValue`

The docs in `bflatn_to.rs` claim that we'll panic
if the `ProductValue` passed to `write_row_to_pages`
does not conform to the type.

Prior to this commit, we would panic if some element of the `val`
had a different type constructor than the corresponding element of the `ty`,
but for `ProductValue`, we would not panic if the value had more or fewer elements
than the type.
When the value had more elements, this was merely a logic bug;
we'd drop the tail elements and go about our lives.
When the value had fewer elements, this invoked UB;
we'd leave the remainder of the row in the page uninit,
and later access uninitialized memory e.g. when hashing the row.

This commit adds a check that in `write_product`
that the `val` and `ty` have the same length.
If they do not, this commit panics.
This will need to be revised to return an `Err`, as spacetimedb-core has tests
that attempting to insert an ill-typed value returns an `Err` rather than panicking.

* Error handling, runtime schema mutations

This commit makes two sorts of changes:

- It begins filling in the error-handling story of the new mem-arch
  by defining a few new error enums in the `spacetimedb-table` crate,
  and converting them into `TableError` in `spacetimedb-core`.
- It implements `drop_table`, `rename_table` and friends to mutate a schema at runtime.
  These operations are broken in master in that they violate transactionality,
  and they're broken here too.

* Re-enable metrics

* Resolve some TODOs, incl. error type in `page.rs`

* Fixes for Tyler's comments

- Restrict some `pub` to `pub(crate)`
- Rename `MutTxId::delete_by_rel` to `delete_by_row_value`, since it deletes only one row.
- `relations` -> `relation`.
- Rename `ignore_duplicate_insert` -> `ignore_duplicate_insert_error` and add docs.
- nix `with_table_*` methods in favor of imperative `get_table_*`.

* Note correctness of `state_view::Iter` and add test in `relational_db`.

Tyler had concerns about the behavior of `state_view::Iter`
w.r.t. the edge case of a committed-deleted-inserted row,
since the new datastore maintains the invariant that
the committed and inserted tables are disjoint,
where the previous datastore did not.

This commit adds comments describing that invariant to the iterator,
and adds a test in `relational_db` that the external semantics are preserved.

* Add notes for future MVCC implementors re: disjoint committed and insert tables

Tyler convinced me via video call that, in MVCC, it is no longer valid
to treat the committed and insert tables as disjoint,
i.e. to elide inserts of already-present rows.

Doing so is still valid, and is more performant, in the current locking model,
so we retain that behavior here.

This commit adds comments to relevant code paths so that, when we move to MVCC,
we do not forget to remove the optimization.

---------

Co-authored-by: Phoebe Goldman <phoebe@goldman-tribe.org>