`ScanIterByColRange`: Avoid unnecessary `ProductValue` allocations (#804)
Prior to this commit, `ScanIterByColRange::next` called `RowRef::to_product_value`
and then `ProductValue::project_not_empty` on every row in the table
to get a key, which it compared to the sought range.
This always allocated at least a `ProductValue`,
even when seeking a range of primitive type like `u64`.
It also unnecessarily deserialized (and potentially allocated)
columns not relevant to the search for rows which did not match the range.

With this commit, we call `RowRef::project_not_empty` directly.
When the sought range is a single column, this avoids allocating a `ProductValue`.
Even when seeking a multi-column range, this avoids deserializing unrelated columns
of non-matching rows.

Non-indexed `IterByColEq` is always going to be slow, so this probably doesn't matter,
but it's a trivial change.