Incremental joins: compile once, run repeatedly. (#938)
* Incremental joins: compile once, run repeatedly.

Well, more like, compile 3 times, run repeatedly, but 3 is approximately 1.

This commit re-writes `IncrementalJoin` to be a re-usable representation
of a query plan for an incremental join,
where before it was a one-off worker for the same.

`IncrementalJoin` stores three copies of the query, compiled for `MemTable * DbTable`,
`DbTable * MemTable` and `MemTable * MemTable`.

Related to this change, `eval_incr` no longer needs an `AuthCtx`,
because we check permissions during query compilation, not execution,
and all query planning is now done ahead of time during `add_subscriber`.
As a result, many callsites, especially in tests, which used to pass an `AuthCtx`
no longer do so.

* `IncrementalJoin`: save `MemTable` headers to avoid recomputing

* Don't include `__op_type` column in incremental joins

Incremental joins never used the `__op_type` column,
as they separate deletes from inserts and eval them as separate queries.

This commit causes incremental joins to no longer include the `__op_type` column
in their `MemTable`s at all,
which simplifies the code and should remove some allocations.

* Remove `pub` on test `run_query` helper.

The module's not `pub`, so it was a pointless qualifier.