Increase sdk tests' timeout to 60 seconds (#956)
We're seeing intermittent failures of the SDK tests in CI due to timeouts.
The timeout is meant only to fail if an expected event never happens;
the SDK tests are not interested in measuring performance at all.

This commit doubles the timeout from 30 to 60 seconds,
in the hopes that we will see fewer false failures.