core: Make `init_module_host` consistent with `update_module_host` (#977)
`spawn_module_host` was changed in #904 to ensure that the supplied
program is indeed the database's current program. There are, however,
uses which would call `spawn_module_host`, followed by either
`init_database` or `update_database` on the returned `ModuleHost`.

This would fail because the condition is not met.

While we cannot prevent misuse, this patch documents the various
lifecycle methods, and adjusts the signature of `init_module_host` to be
(somewhat) consistent with `update_module_host`. Namely, it will
optionally return the result of the `init` reducer call.