fix(1173): Record wait time for all reducers (#1174)
Fixes #1173.

Previously we were only recording this metric for scheduled reducers.
We were also recording it before we acquired access to the module instance.
Now we record it for all reducers after we acquire access to the module instance.

This patch also removes max wait time since the histogram should suffice.