Refactor and extract `call_reducer` and `update_database` machinery for reuse by V8 (#3190)
# Description of Changes

Refactors the machinery of call_reducer and update_database to
a) have smaller pieces that make up the whole so that the code becomes
clearer
b) extract all the VM-independent stuff so that it can be reused by V8
modules.

This is best reviewed commit by commit.

# API and ABI breaking changes

None.

# Expected complexity level and risk

2, it's an important place, but this is doing just code motion.

# Testing

No semantic changes, just code motion.